Refactored of the dashboard views: split them in 4 module base, officer, customer and assessor
